finished project3

+++ project_2_3_4_5/Core.h
@@ -24,8 +24,7 @@ typedef struct Core
     // What else you need? Data memory? Register file?
     Instruction_Memory *instr_mem;
    
-    Byte data_mem[1024]; // data memory
-    // uint64_t data_mem[32]; // data memory
+    int64_t data_mem[1024]; // data memory
   
     Register reg_file[32]; // register file.
